move 'spinner' code into AppController, move 'allow/stopInteraction' code to iosish/RootController
